MNT Cleanup include paths and normalize the meson.build file a bit to try and bring things back under control...
