+++ workflows/maestro/src/lwm/job.cc
@@ -1,4 +1,4 @@
-#include "job.hh"
+#include "lwm/job.hh"
 
 #include <stdexcept>
 #include <string>
