+++ workflows/maestro/src/parallel/threadpool.cc
@@ -1,4 +1,4 @@
-#include "threadpool.hh"
+#include "parallel/threadpool.hh"
 
 #include "spdlog/spdlog.h"
 #include "spdlog/sinks/stdout_color_sinks.h"
