+++ workflows/maestro/src/server/handler.cc
@@ -1,4 +1,4 @@
-#include "handler.hh"
+#include "server/handler.hh"
 
 #include "rapidjson/document.h"
 #include "rapidjson/stringbuffer.h"
