+++ workflows/maestro/src/server/http.cc
@@ -1,4 +1,4 @@
-#include "http.hh"
+#include "server/http.hh"
 
 #include <algorithm>
 #include <cctype>
