+++ workflows/maestro/tests/cpp/test_http.cc
@@ -1,4 +1,4 @@
-#include "http.hh"
+#include "server/http.hh"
 
 #include <gtest/gtest.h>
